// SDKs/twilio.com/twilio_sync_v1/1.37.3/typescript/src/sdk/models/operations/listsynclist.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import * as shared from "../shared";

export const ListSyncListServerList = [
	"https://sync.twilio.com",
] as const;


export class ListSyncListPathParams extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=ServiceSid" })
  serviceSid: string;
}


export class ListSyncListQueryParams extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=PageSize" })
  pageSize?: number;
}


export class ListSyncListSecurity extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "security, scheme=true;type=http;subtype=basic" })
  accountSidAuthToken: shared.SchemeAccountSidAuthToken;
}


export class ListSyncListListSyncListResponseMeta extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=first_page_url" })
  firstPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=key" })
  key?: string;

  @SpeakeasyMetadata({ data: "json, name=next_page_url" })
  nextPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=page" })
  page?: number;

  @SpeakeasyMetadata({ data: "json, name=page_size" })
  pageSize?: number;

  @SpeakeasyMetadata({ data: "json, name=previous_page_url" })
  previousPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}


export class ListSyncListListSyncListResponse extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=lists", elemType: shared.SyncV1ServiceSyncList })
  lists?: shared.SyncV1ServiceSyncList[];

  @SpeakeasyMetadata({ data: "json, name=meta" })
  meta?: ListSyncListListSyncListResponseMeta;
}


export class ListSyncListRequest extends SpeakeasyBase {
  @SpeakeasyMetadata()
  serverUrl?: string;

  @SpeakeasyMetadata()
  pathParams: ListSyncListPathParams;

  @SpeakeasyMetadata()
  queryParams: ListSyncListQueryParams;

  @SpeakeasyMetadata()
  security: ListSyncListSecurity;
}


export class ListSyncListResponse extends SpeakeasyBase {
  @SpeakeasyMetadata()
  contentType: string;

  @SpeakeasyMetadata()
  listSyncListResponse?: ListSyncListListSyncListResponse;

  @SpeakeasyMetadata()
  statusCode: number;
}
